Add support for set equality
