+++ packages/earljs/src/isEqual/index.ts
@@ -1,4 +1,5 @@
 export * from './EqualityOptions'
 export { CanonicalType, getCanonicalType } from './getCanonicalType'
+export { getKeys } from './getKeys'
 export { isEqual } from './isEqual'
-export { getKeys, isEqualUnknown } from './isEqualUnknown'
+export { isEqualUnknown } from './isEqualUnknown'
