+++ packages/earljs/src/isEqual/isEqualSet.ts
@@ -1,0 +1,13 @@
+export function isEqualSet(value: Set<unknown>, other: Set<unknown>) {
+  if (value.size !== other.size) {
+    return false
+  }
+
+  for (const item of other) {
+    if (!value.has(item)) {
+      return false
+    }
+  }
+
+  return true
+}
